move rich text editor
