+++ src/app/test-pages/rich-text-editor/page.js
@@ -1,4 +1,4 @@
-import RichTextEditorDemo from "@/components/RichTextEditor/demo";
+import RichTextEditorDemo from "@/components/Admin/RichTextEditor/demo";
 
 export default function TestRichTextPage() {
   return <RichTextEditorDemo />;
