+++ src/components/Admin/RichTextEditor/RichTextEditor.module.scss
@@ -1,5 +1,5 @@
 /* Import Typography styles */
-@import "../Typography/Typography.module.scss";
+@import "../../Typography/Typography.module.scss";
 
 /* RichTextEditor styles that use Typography component classes */
 .richTextEditor {
